The actual middleware.
This middleware will look at an Authorization HTTP header, and will
check the supplied token on a standard OAuth2 Token Introspection
endpoint for validity.

This means that this middleware will allow API's to validate with
a12n-server, but also any other compliant OAuth2 server that supports
RFC7662.

If the client did have a valid token, this middleware will store the
results of the the introspection request in Curveballs'
ctx.state.oauth2.