Always check authentication header if it's there.
The OAuth2 package has an option to *not* check authentication for
certain public paths. This previously was called 'whitelist', but is
renamed to 'publicPrefixes' for future versions.

If a user hit those paths, no auhtentication would be done.

The change this PR makes is that auth is checked if you are hitting a
public path, and you specified an Authorization header.

The benefit is that even for public paths authenticated clients might
send auth headers, and it might be useful for applications to know
whether or not a user was authenticated or not.